+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.4.0 (pending)
 
 * renamed *.log to *.protocol
-* renamed .enabled to freigeschlatet.selection
+* renamed .enabled to freigeschaltet.selection
 * update parent 1.5.0 to 1.5.1
 * update spring-boot 2.4.4 to 2.5.5
 * update bootstrap 4.6.0 to 5.1.1